requirements.txt
django-picklefield
git+https://github.com/if-then-fund/django-html-emailer
